-Fix the client counting all items from all games when scanning for recieved items and songs -Add a command to check your fame point total and goal

+++ Data/CommonData.cs
@@ -66,7 +66,7 @@ namespace YargArchipelagoClient.Data
             public bool WasGoldStar;
             public bool FC;
             public int Score;
-            public double Percentage;
+            public float Percentage;
         }
 
         public enum SupportedDifficulty
